SKL Refactor small data code to base class. Begin XAS.
